Scaled MCP server/requirements.txt
Authlib==1.6.4
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.19.0
Werkzeug==3.1.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
cyclopts==3.24.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.12.3
grpcio-tools==1.68.0
grpcio==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx-sse==0.4.1
httpx==0.28.1
idna==3.10
isodate==0.7.2
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
mcp==1.14.1
mdurl==0.1.2
more-itertools==10.8.0
nest-asyncio==1.6.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
parse==1.20.2
pathable==0.4.4
pycparser==2.23
pydantic-settings==2.10.1
pydantic==2.11.9
pydantic_core==2.33.2
pyperclip==1.10.0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich-rst==1.3.1
rich==14.1.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0